Remove env Refactor move image_type and job_type into a single file
