+++ src/constants.js
@@ -1,0 +1,22 @@
+const IMAGE_TYPE = {
+  SELFIE_IMAGE_FILE: 0,
+  ID_CARD_IMAGE_FILE: 1,
+  SELFIE_IMAGE_BASE64: 2,
+  ID_CARD_IMAGE_BASE64: 3,
+  LIVENESS_IMAGE_FILE: 4,
+  ID_CARD_BACK_IMAGE_FILE: 5,
+  LIVENESS_IMAGE_BASE64: 6,
+  ID_CARD_BACK_IMAGE_BASE64: 7,
+};
+
+const JOB_TYPE = {
+  BIOMETRIC_KYC: 1,
+  SMART_SELFIE_REGISTRATION: 2,
+  SMART_SELFIE_AUTHENTICATION: 4,
+  BASIC_KYC: 5,
+  ENHANCED_KYC: 5,
+  DOCUMENT_VERIFICATION: 6,
+  BUSINESS_VERIFICATION: 7,
+};
+
+module.exports = { IMAGE_TYPE, JOB_TYPE, };
