--- src/constants/env.js
@@ -1,10 +1,0 @@
-const ENV = {
-  TEST: 1,
-  LIVE: 2,
-  SID_SERVER_MAPPING: {
-    0: 'testapi.smileidentity.com/v1',
-    1: 'api.smileidentity.com/v1',
-  },
-};
-
-module.exports = ENV;
