--- src/constants/image_type.js
@@ -1,12 +1,0 @@
-const IMAGE_TYPE = {
-  SELFIE_IMAGE_FILE: 0,
-  ID_CARD_IMAGE_FILE: 1,
-  SELFIE_IMAGE_BASE64: 2,
-  ID_CARD_IMAGE_BASE64: 3,
-  LIVENESS_IMAGE_FILE: 4,
-  ID_CARD_BACK_IMAGE_FILE: 5,
-  LIVENESS_IMAGE_BASE64: 6,
-  ID_CARD_BACK_IMAGE_BASE64: 7,
-};
-
-module.exports = IMAGE_TYPE;
