--- src/constants/job_type.js
@@ -1,11 +1,0 @@
-const JOB_TYPE = {
-  BIOMETRIC_KYC: 1,
-  SMART_SELFIE_REGISTRATION: 2,
-  SMART_SELFIE_AUTHENTICATION: 4,
-  BASIC_KYC: 5,
-  ENHANCED_KYC: 5,
-  DOCUMENT_VERIFICATION: 6,
-  BUSINESS_VERIFICATION: 7,
-};
-
-module.exports = JOB_TYPE;
