+++ test/index.test.js
@@ -6,15 +6,14 @@ describe('smile-identity-core', () => {
     assert.equal(typeof smileIdentityCore, 'object');
   });
 
-  it('should export four classes and three objects', () => {
+  it('should export four classes and two objects', () => {
     assert.equal(typeof smileIdentityCore.Signature, 'function');
     assert.equal(typeof smileIdentityCore.Utilities, 'function');
     assert.equal(typeof smileIdentityCore.WebApi, 'function');
     assert.equal(typeof smileIdentityCore.IDApi, 'function');
-    assert.equal(typeof smileIdentityCore.ENV, 'object');
     assert.equal(typeof smileIdentityCore.IMAGE_TYPE, 'object');
     assert.equal(typeof smileIdentityCore.JOB_TYPE, 'object');
-    assert.equal(Object.keys(smileIdentityCore).length, 7);
+    assert.equal(Object.keys(smileIdentityCore).length, 6);
   });
 
   it('should run in node and not in the browser', () => {
